Move src/eva/models/versioning.py to src/eva/metrics/
